+++ source/renderer/app/containers/wallet/WalletSummaryPage.js
@@ -41,7 +41,7 @@ export default class WalletSummaryPage extends Component<Props> {
 
   render() {
     const { intl } = this.context;
-    const { app, wallets, transactions } = this.props.stores;
+    const { app, wallets, transactions, profile } = this.props.stores;
     const {
       openExternalLink,
       environment: { network },
@@ -55,6 +55,7 @@ export default class WalletSummaryPage extends Component<Props> {
     } = transactions;
     const { isActiveWalletRestoring } = wallets;
     const wallet = wallets.active;
+    const { currentTimeFormat, currentDateFormat } = profile;
     // Guard against potential null values
     if (!wallet)
       throw new Error('Active wallet required for WalletSummaryPage.');
@@ -84,6 +85,8 @@ export default class WalletSummaryPage extends Component<Props> {
           onOpenExternalLink={openExternalLink}
           onShowMoreTransactions={this.handleShowMoreTransaction}
           totalAvailable={totalAvailable}
+          currentTimeFormat={currentTimeFormat}
+          currentDateFormat={currentDateFormat}
         />
       );
     } else if (!hasAny) {