+++ source/renderer/app/containers/wallet/dialogs/paper-wallet-certificate/InstructionsDialogContainer.js
@@ -1,28 +1,16 @@
 // @flow
 import React, { Component } from 'react';
 import { observer, inject } from 'mobx-react';
-import { defineMessages, intlShape } from 'react-intl';
 import moment from 'moment';
 import InstructionsDialog from '../../../../components/wallet/paper-wallet-certificate/InstructionsDialog';
 import type { InjectedDialogContainerProps } from '../../../../types/injectedPropsType';
 import { generateFileNameWithTimestamp } from '../../../../../../common/utils/files';
-
-const messages = defineMessages({
-  timestamp: {
-    id: 'paper.wallet.create.certificate.instructions.dialog.timestamp',
-    defaultMessage: '!!!MMMM D, YYYY - h:mm A',
-    description: 'Timestamp for the Paper Wallet PDF content.',
-  },
-});
 
 type Props = InjectedDialogContainerProps;
 
 @inject('stores', 'actions')
 @observer
 export default class InstructionsDialogContainer extends Component<Props> {
-  static contextTypes = {
-    intl: intlShape.isRequired,
-  };
   static defaultProps = {
     actions: null,
     stores: null,
@@ -31,10 +19,14 @@ export default class InstructionsDialogContainer extends Component<Props> {
   };
 
   onPrint = () => {
-    const { intl } = this.context;
+    const {
+      currentDateFormat,
+      currentTimeFormatShort,
+    } = this.props.stores.profile;
     const date = moment();
-    const localizedTimestampFormat = intl.formatMessage(messages.timestamp);
-    const timestamp = moment(date).format(localizedTimestampFormat);
+    const formattedDate = date.format(currentDateFormat);
+    const formattedTime = date.format(currentTimeFormatShort);
+    const timestamp = `${formattedDate} - ${formattedTime}`;
 
     const name = generateFileNameWithTimestamp({
       prefix: 'paper-wallet-certificate',