+++ source/renderer/app/stores/AppStore.js
@@ -123,7 +123,7 @@ export default class AppStore extends Store {
 
   @computed get isSetupPage(): boolean {
     return (
-      this.currentRoute === ROUTES.PROFILE.LANGUAGE_SELECTION ||
+      this.currentRoute === ROUTES.PROFILE.INITIAL_SETTINGS ||
       this.currentRoute === ROUTES.PROFILE.TERMS_OF_USE
     );
   }