+++ storybook/stories/news/NewsFeed.stories.js
@@ -3,10 +3,12 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { action } from '@storybook/addon-actions';
-import { boolean } from '@storybook/addon-knobs';
+import { boolean, select, withKnobs } from '@storybook/addon-knobs';
 import StoryDecorator from '../_support/StoryDecorator';
 import NewsFeed from '../../../source/renderer/app/components/news/NewsFeed';
 import News from '../../../source/renderer/app/domains/News';
+import { dateOptions } from '../_support/profileSettings';
+import { DATE_ENGLISH_OPTIONS } from '../../../source/renderer/app/config/profileConfig';
 
 const news = [
   new News.News({
@@ -113,8 +115,8 @@ const news = [
 const newsCollection = new News.NewsCollection(news);
 
 storiesOf('News|NewsFeed', module)
-  .addDecorator(story => (
-    <StoryDecorator>{story(newsCollection)}</StoryDecorator>
+  .addDecorator((story, context) => (
+    <StoryDecorator>{withKnobs(story, context)}</StoryDecorator>
   ))
 
   // ====== Stories ======
@@ -132,6 +134,7 @@ storiesOf('News|NewsFeed', module)
         onOpenExternalLink={action('onOpenExternalLink')}
         onOpenAlert={action('onOpenAlert')}
         onProceedNewsAction={action('onOpenExternalLink')}
+        currentDateFormat=" "
       />
     </div>
   ))
@@ -149,6 +152,7 @@ storiesOf('News|NewsFeed', module)
         onOpenExternalLink={action('onOpenExternalLink')}
         onOpenAlert={action('onOpenAlert')}
         onProceedNewsAction={action('onOpenExternalLink')}
+        currentDateFormat=" "
       />
     </div>
   ))
@@ -166,6 +170,11 @@ storiesOf('News|NewsFeed', module)
         onOpenExternalLink={action('onOpenExternalLink')}
         onOpenAlert={action('onOpenAlert')}
         onProceedNewsAction={action('onOpenExternalLink')}
+        currentDateFormat={select(
+          'currentDateFormat',
+          dateOptions,
+          DATE_ENGLISH_OPTIONS[0].value
+        )}
       />
     </div>
   ));